backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.7.0
groq==0.15.0
pytesseract==0.3.13
Pillow==11.0.0
PyMuPDF==1.25.1
pdfplumber==0.11.4
python-docx==1.1.2
aiofiles==24.1.0
jinja2==3.1.4
fpdf2==2.8.2
httpx==0.28.1
python-dotenv==1.0.1
